requirements.txt
pyautogui
pillow
tkcap
pydicom
img2pdf
matplotlib
pandas
tensorflow
python-xlib